Flowers added

+++ src/Blume.java
@@ -1,5 +1,5 @@
 public class Blume {
-    float preis;
+    double preis;
     int groesseInCm;
     boolean istLebendig;
     int bevorzugteTemperatur;
@@ -10,7 +10,7 @@ public class Blume {
         this.istLebendig = true;
     }
 
-    public Blume(final float preis, final int groesseInCm, final boolean istLebendig, final int bevorzugteTemperatur, final boolean brauchtTageslicht) {
+    public Blume(final double preis, final int groesseInCm, final boolean istLebendig, final int bevorzugteTemperatur, final boolean brauchtTageslicht) {
         this.preis = preis;
         this.groesseInCm = groesseInCm;
         this.istLebendig = istLebendig;
