+++ src/BlumeMitSpitznameEinheimisch.java
@@ -2,7 +2,7 @@ public class BlumeMitSpitznameEinheimisch extends BlumeMitSpitzname {
 
     boolean einheimisch;
 
-    BlumeMitSpitznameEinheimisch(final boolean einheimisch, final String spitzname, final float preis, final int groesseInCm, final boolean istLebendig, final int bevorzugteTemperatur, final boolean brauchtTageslicht) {
+    BlumeMitSpitznameEinheimisch(final boolean einheimisch, final String spitzname, final double preis, final int groesseInCm, final boolean istLebendig, final int bevorzugteTemperatur, final boolean brauchtTageslicht) {
         super(spitzname, preis, groesseInCm, istLebendig, bevorzugteTemperatur, brauchtTageslicht);
         this.einheimisch = einheimisch;
     }
